Exception handling

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -1,16 +1,20 @@
 package com.devsuperior.dscommerce.services;
 
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
+import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -20,7 +24,8 @@ public class ProductService {
 	
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {		 
-		 Product product = repository.findById(id).get();
+		 Product product = repository.findById(id).orElseThrow(
+				 () -> new ResourceNotFoundException("Recurso não encontrado"));
 		 return new ProductDTO(product);	 
 		 
 	}
@@ -44,19 +49,33 @@ public class ProductService {
 	}
 	
 	@Transactional
-	public ProductDTO update(Long id, ProductDTO dto) {			
-		Product entity = repository.getReferenceById(id);
-		copyDtotoEntity(dto, entity);		
-		entity = repository.save(entity);
+	public ProductDTO update(Long id, ProductDTO dto) {		
 		
-		return new ProductDTO(entity);
+		try {
+			Product entity = repository.getReferenceById(id);
+			copyDtotoEntity(dto, entity);		
+			entity = repository.save(entity);
+			
+			return new ProductDTO(entity);
+			
+		}catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		
 		 
 	}
 	
-	@Transactional
-	public void delete(Long id) {		 
-		repository.deleteById(id); 
-		 
+	@Transactional(propagation = Propagation.SUPPORTS)
+	public void delete(Long id) {
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+	        	repository.deleteById(id);    		
+		}
+	    	catch (DataIntegrityViolationException e) {
+	        	throw new DatabaseException("Falha de integridade referencial");
+	   	}
 	}
 
 	private void copyDtotoEntity(ProductDTO dto, Product entity) {
